move lib/replace.js -> index.js
